capstone presentation

+++ Arc.Device/src/components/Led.h
@@ -14,7 +14,7 @@ namespace arc
 				Led(PinName pin);
 
 				void solid();
-				void pulse(int pulseCount = 1);
+				void pulse(uint8_t pulseCount = 1);
 				void blink(bool fast = false);
 				void off();
 			private:
@@ -23,8 +23,8 @@ namespace arc
 				EventQueue queue;
 				Thread th;
 
-				static const int pulseDuration = 150;
-				Event<void(int)> pulseEv;
+				static const uint8_t pulseDuration = 150;
+				Event<void(uint8_t)> pulseEv;
 				int pulseId;
 
 				int blinkDuration;
@@ -33,7 +33,7 @@ namespace arc
 				int blinkOnId;
 				int blinkOffId;
 
-				void do_pulse(int pulseCount);
+				void do_pulse(uint8_t pulseCount);
 				void do_blinkOn();
 				void do_blinkOff();
 				void cancelTasks();
